+++ fns/add/coin_send.php
@@ -4,20 +4,21 @@ if(role(['permissions' => ['coins' => 'allow_conversion']])) {
     if (isset($data['recipient_user_id'], $data['coins_to_send']) && is_numeric($data['coins_to_send'])) {
         $sender_user_id = (int)Registry::load('current_user')->id;
         $recipient_user_id = $data['recipient_user_id'];
-        $coins_to_send = (int) $data['coins_to_send'];
+        $coins_to_send = (float) $data['coins_to_send'];
         if(role(['permissions' => ['groups' => 'super_privileges']]))
         {
-            $minimum_value_to_send_coins = Registry::load('settings')->minimum_conversion_value;
-            $maximum_value_to_send_coins = Registry::load('settings')->maximum_conversion_value;
+            $minimum_value_to_send_coins = (float)Registry::load('settings')->minimum_conversion_value;
+            $maximum_value_to_send_coins = (float)Registry::load('settings')->maximum_conversion_value;
         }
         else
         {
-            $minimum_value_to_send_coins = role(['find' => 'minimum_value_to_convert_coins']);
-            $maximum_value_to_send_coins = role(['find' => 'maximum_value_to_convert_coins']);
+            $minimum_value_to_send_coins = (float)role(['find' => 'minimum_value_to_convert_coins']);
+            $maximum_value_to_send_coins = (float)role(['find' => 'maximum_value_to_convert_coins']);
         }
+
         $available_balance_to_send_coins = role(['find' => 'available_balance_to_convert_coins']);
         if($minimum_value_to_send_coins > $coins_to_send)
-        {
+        {   
             $result['success'] = false;
             $result['error_message'] = 'You can\' not send less than '.$minimum_value_to_send_coins.' coins';
             return;
@@ -46,7 +47,7 @@ if(role(['permissions' => ['coins' => 'allow_conversion']])) {
                 // Update recipient's balance
                 DB::connect()->update(
                     "user_coins",
-                    ["coins_balance" => ((int)$recipient_coin_balance[0] + $coins_to_send)],
+                    ["coins_balance" => ((float)$recipient_coin_balance[0] + $coins_to_send)],
                     ["user_id" => $recipient_user_id]
                 );
             } else {
@@ -57,7 +58,7 @@ if(role(['permissions' => ['coins' => 'allow_conversion']])) {
                 );
             }
             // Update recipient's balance
-            DB::connect()->update("user_coins", ["coins_balance" => ((int)$sender_balance[0] - $coins_to_send)], ["user_id" => $sender_user_id]);
+            DB::connect()->update("user_coins", ["coins_balance" => ((float)$sender_balance[0] - $coins_to_send)], ["user_id" => $sender_user_id]);
     
             // Record the transaction
             DB::connect()->insert("coins", [
